fix try

+++ Framework/src/RepositoryBenchmark.cxx
@@ -192,7 +192,7 @@ TH1* RepositoryBenchmark::createHisto(uint64_t sizeObjects, string name, bool ol
     mo->setIsOwner(true);
     mMyObjects.push_back(mo);
   }
-  for (uint64_t i = 0 ; i < nbSmallObjects ; i++) {
+  for (int i = 0 ; i < nbSmallObjects ; i++) {
     TH1* histo = createHisto(mSizeObjects, mObjectName + "_small_" + to_string(i), objectsSizeOldBehaviour);
     shared_ptr<MonitorObject> mo = make_shared<MonitorObject>(histo, mTaskName, "BMK");
     mo->setIsOwner(true);
